FastAPI/requirement.txt.txt
fastapi
uvicorn
pydantic
PyPDF2
python-docx
pillow
pytesseract
